% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{as_interaction}
\alias{as_interaction}
\alias{as.interaction}
\alias{as_interaction.default}
\alias{as_interaction.interaction}
\title{as.interaction}
\usage{
as_interaction(object)

as.interaction(object)

\method{as_interaction}{default}(object)

\method{as_interaction}{interaction}(object)
}
\description{
\code{as.interaction} transforms a vector in an interactions matrix object defining the positive and negative interactions between populations.
The interaction is not assumed symetrical, so a population can have a different effect on a population than this latest has on the first one.
The interaction of a population on itself is equal to 0. If no interaction vector is given, the default values are 0.
}
\examples{
  as.interaction(c(0.2,-0.5, 0.1,0.2,0.3,0.8))

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
