% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_population.R
\name{as_population}
\alias{as_population}
\alias{as_population.default}
\alias{as_population.population}
\alias{as_population.list}
\title{as_population}
\usage{
as_population(object)

\method{as_population}{default}(object)

\method{as_population}{population}(object)

\method{as_population}{list}(object)
}
\arguments{
\item{object}{a list of populations objects (information on the population)}
}
\value{
If \code{'object'} is a list, the function will test if it match with the
 the requirements of a class \code{'population'} object.
 If \code{'object'} is not a list an error message will be give.
}
\description{
The fonction \code{'as_population'} takes as argument a list
containing the population caracteristics (population,
number at initial time, growth rates.
And check if is a class population and  return a list duable by \code{BeePODYNA} package
}
\examples{
 obj = list(label ="pop.test",size= 100,growth_rate = 0.1,capacity = 500)
 as_population(obj)

}
\author{
Martial <leroy.martial@etu.univ-grenoble-alpes.fr>
}
