% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{beepod_sweden}
\alias{beepod_sweden}
\title{Data Sweden : 3 populations in iteraction in Sweden}
\format{A Beepodyna object of 1 community with 3 populations:
\describe{
  \item{Sweden}{the label of the community}
  \item{interaction}{the matrix of interaction of the 3 populations, here 0 interaction}
 \item{populations}{the list of two populations}
   \itemize{
   \item{red_fox : the population of red, a list of : growth rate, biotic capacity,initial size}
   \item{gouse : the population of gouse, a list of : growth rate, biotic capacity,initial size}
   \item{vole : the population of vole, a list of : growth rate, biotic capacity,initial size}
  }

}}
\source{
\url{https://www.nature.com/scitable/knowledge/library/dynamics-of-predation-13229468/}
}
\usage{
beepod_sweden
}
\description{
3 populations in interactions have been reported between red foxes its major prey voles and
most rarely grouve if voles are not available. Here is a classical case of bottom-up control.
}
\details{
Red fox population is characterized by low growth rate and a little capacity.
Gouse population is characterized by a medium growth rate and a medium capacity.
Vole population is characterized by a high growth rate and a high capacity.
}
\author{
Leroy Martial <martial.leroy@etu.univ-grenoble-alpes.fr>
}
\keyword{datasets}
