% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{capacity_population}
\alias{capacity_population}
\title{Returns the capacity of the population}
\usage{
capacity_population(pop)
}
\arguments{
\item{pop}{is an object of the class 'population'}
}
\value{
returns the capacity of the population.
}
\description{
Returns the capacity of the population
}
\examples{
    pop=population("Pop_1",42,2,200)
    capacity_population(pop)

}
\author{
Nicolas BARTALUCCI
}
