% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{grate_population}
\alias{grate_population}
\title{Returns the growth rate of the population}
\usage{
grate_population(pop)
}
\arguments{
\item{pop}{is an object of the class 'population'}
}
\value{
returns the growth rate of the population.
}
\description{
Returns the growth rate of the population
}
\examples{
    pop<-population("Pop_1",42,2,200)
    grate_population(pop)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
