% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hudson}
\alias{hudson}
\title{hudson}
\format{An object of class \code{community} of length 2.}
\usage{
hudson
}
\description{
Community hudson
}
\examples{
data(hudson)
plot(hudson)
summary(hudson)
}
\seealso{
\code{\link[BeePODYNA]{beepod_hudson}} for description of the dataset.
}
\keyword{datasets}
