% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{is.interaction}
\alias{is.interaction}
\title{is.interaction}
\usage{
is.interaction(x)
}
\arguments{
\item{x}{is an \code{R} object.}
}
\description{
\code{is.interaction} returns TRUE if x is an interaction matrix of the class 'interaction'.
It returns FALSE otherwise.
To create such an object you can use the \code{\link[BeePODYNA]{mat_interaction}} function.
}
\examples{
  is.interaction(3)
  x=mat_interaction(nb_pop=3, interactions=c(0.2,-0.5, 0.1,0.2,0.3,0.8))
  is.interaction(x)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
