% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_population.R
\name{is.population}
\alias{is.population}
\title{Is.population
The function determinines if the object is a population or not}
\usage{
is.population(x)
}
\arguments{
\item{x}{the object which must be a population to validate the condition}
}
\value{
a logical "TRUE" or "FALSE"
}
\description{
Is.population
The function determinines if the object is a population or not
}
\author{
Cresciense Lecaude
}
