% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_population.R
\name{is_population}
\alias{is_population}
\title{Tests if an object is belonging the class \code{population}.}
\usage{
is_population(x)
}
\arguments{
\item{x}{the object which must be a population to validate the condition}
}
\value{
a logical "TRUE" or "FALSE"
}
\description{
The function determinines if the object is a population or not
}
\examples{

  p <- population('example', 2, 0.5, 10)
  is_population(p) # Returns TRUE
  is_population(4) # Returns FALSE


}
\author{
Cresciense Lecaude
}
