% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{length.community}
\alias{length.community}
\title{Returns the number of species described in the community}
\usage{
\method{length}{community}(x)
}
\arguments{
\item{x}{the considered community}
}
\description{
Returns the number of species described in the community
}
\examples{
   hare = population("hirsuta",30,2,80)
   lynx = population("daonensis",4,1.2,60)

   hudson = community('hudson',hare,lynx)
   length(hudson)

}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
