% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{mat_interaction}
\alias{mat_interaction}
\title{'mat-interaction' creates an interactions matrix object}
\usage{
mat_interaction(nb_pop, interactions = rep(0, nb_pop * (nb_pop - 1)))
}
\arguments{
\item{nb_pop}{is the number of populations in the model.}

\item{interactions}{is a vector of length nb_pop*(nb_pop-1) giving the interaction of each population on the other one.
Each interaction is a decimal ranging between -1 and 1. A positive value means a positive impact (facilitation) while a negative value means a negative impact (predation, competition).
The vector starts with the vector of the interactions of the first population on the other ones (ranging from 2 to the last one, not including itself),
and then the vector of the interactions of the second population on the others...}
}
\description{
Creates an interactions matrix object defining the positive and negative interactions between populations.
The interaction is not assumed symetrical, so a population can have a different effect on a population than this latest has on the first one.
The interaction of a population on itself is equal to 0.
If no interaction vector is given, the default values are 0.
}
\examples{
  mat_interaction(nb_pop=3, interactions=c(0.2,-0.5, 0.1,0.2,0.3,0.8))
  mat_interaction(nb_pop=2)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
