% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{names_population}
\alias{names_population}
\title{Returns the name of the population}
\usage{
names_population(pop)
}
\arguments{
\item{pop}{is an object of the class 'population'}
}
\value{
returns the name of the population.
}
\description{
Returns the name of the population
}
\examples{
    pop=population("test",42,2,200)
    names.population(pop)

}
\author{
Nicolas BARTALUCCI
}
