% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_population}
\alias{plot_population}
\alias{plot_population.population}
\alias{plot.population}
\title{plot.population}
\usage{
plot_population(x, ...)

\method{plot_population}{population}(pop, ...)

\method{plot}{population}(pop, ...)
}
\arguments{
\item{x}{a population object}

\item{...}{every argument for the \code{\link[graphics]{plot}} function such as graphical parameters for lines. Does NOT apply for lines and capacity ablines.
\itemize{
  \item capacity.line TRUE/FALSE if a dashed line must be drawned for each populations
  \item log.time TRUE/FALSE if time axis (x axis) must be logarithmic
  \item capacity_lty set to 3 by default
  \item type for the size line, set to 'b' by default
  \item xlab set to 'Time' by default
  \item ybal set to 'Size' by default
  \item main set to use the pop label by default
  \item capacity_line set to TRUE by default
  }}
}
\description{
plot one or multiple given populations
}
\examples{
\dontrun{
hirsu = population("hirsuta",20,1,100)
daonen = population("daonensis",30,1.2,100)

daonen[[2]]=c(30,34)
daonen[[3]]=c(0,5)

liste = list(hirsu,daonen)
plot_population(list = liste, log.pop = T, color = c(3,2), xlab = "Time")
}

}
\seealso{
\code{\link[graphics]{plot}} and \code{\link[graphics]{par}} for graphical parameter that were omitted on this documentation
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
