% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_n_pop_dynamic}
\alias{simulate_n_pop_dynamic}
\title{Compute n times steps of evolution of the populations contained in a beepodyna object}
\usage{
simulate_n_pop_dynamic(beepodyna, n)
}
\arguments{
\item{beepodyna}{an object of class beepodyna}

\item{n}{The number of times steps to be computed}
}
\value{
A beepodyna object, every populations in it will be incremented with n more size values (calculated with the functions and interaction object) and time values
}
\description{
Compute n times steps of evolution of the populations contained in a beepodyna object
}
\examples{
   data(hudson)
   beepop=beepodyna(label="model_1",
   community=hudson,
   interactions=interactions(2),
   list(exponential_growth,exponential_growth),
   verbose = FALSE
    )
    bepop10=simulate_n_pop_dynamic(beepop,10)

}
