% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_onestep}
\alias{simulate_onestep}
\title{Compute one time-step of  evolution of the populations contained in a beepodyna object}
\usage{
simulate_onestep(beepodyna)
}
\arguments{
\item{beepodyna}{an object of class beepodyna}
}
\value{
A beepodyna object, every populations in it will be incremented with one more size value (calculated with the functions and interaction matrix) and time
}
\description{
Compute one time-step of  evolution of the populations contained in a beepodyna object
}
\examples{
  data(hudson)
  beepop=beepodyna(label="model_1",
  community=hudson,
  interactions=interactions(2),
  list(exponential_growth,exponential_growth),
  verbose = FALSE
  )
  bepop1=simulate_onestep(beepop)

}
