% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_population.R
\name{summary.community}
\alias{summary.community}
\title{Communities summaries and trends
#' @author Julia Guerra <jguerra_1995@hotmail.es>}
\usage{
\method{summary}{community}(object, ...,
  d.print = as.numeric(getOption("digits")))
}
\arguments{
\item{object}{an object of class "community"}

\item{...}{not implemented yet}

\item{d.print}{the number of decimal digits to be printed in statistics}
}
\value{
A list containing calculated summaries for a community. See 'examples'.
}
\description{
This function will allow the user to easily read
trends and characteristics (label, population size(s), growth rate(s)...) of one
community-class object.
}
