% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_population.R
\name{summary.population}
\alias{summary.population}
\title{Population summaries and trends
#' @author Julia Guerra <jguerra_1995@hotmail.es>}
\usage{
summary(population, d.print)
}
\arguments{
\item{populations}{an object of class "population" or "community"}

\item{d.print}{the number of decimal digits to be printed in statistics}
}
\value{
A list containing calculated summaries for a population. See 'examples'.
}
\description{
This function will allow the user to easily read
trends and characteristics (label, population size(s), growth rate(s)...) of one or several
population-class objects. When dealing with a list of population-type objects, it will
recursively print the summaries for each one of them.
}
\examples{
#Default usage
pop1 = population("mypop", 1, 0.8, 500)
summary(pop1)





# Several generations and population lists
data(hudson)
pop1 = hudson$hare
sumrs1 = summary(pop1)   #for saving summaries list
sumrsc = summary(hudson)

sumrs1$Rates_of_change        # taking a look into last growth tendencies
sumrsc$hare$Rates_of_change   # same for community object

'

# In case you want to modify digits' printing settings

summary(pop1,11)
summary(pop1,getOption("digits")+2)

}
