% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.population.R
\name{summary_population}
\alias{summary_population}
\title{Population summaries and trends
#' @author Julia Guerra <jguerra_1995@hotmail.es>}
\usage{
summary_population(populations, d.printing = getOption("digits"))
}
\arguments{
\item{d.printing}{the number of decimal digits to be printed in statistics}

\item{pop}{an object of class "population"}
}
\description{
This function will allow the user to easily read
trends and characteristics (label, population size(s), growth rate(s)...) of one or several
population-class objects. When dealing with a list of population-type objects, il will
recursively print the summaries for each one of them.
}
\examples{
# Default usage
pop = population("mypop", 1, 0.8, 500)
summary_population(pop)



# Several generations and population lists
pop1 = population("eva", 1, 0.8, 500)
pop2 = population("teo", 1, 0.7, 850)
pop2$size = c(1,50,750)

pops = community('pops', pop1, pop2)

summary_population(pop2)
summary_population(pops)



# In case you want to modify digits' printing settings

summary_population(pop,11)
summary_population(pop,getOption("digits")+2)


}
