% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sweden}
\alias{sweden}
\title{sweden}
\format{An object of class \code{community} of length 3.}
\usage{
sweden
}
\description{
Community sweden
}
\examples{
data(sweden)
plot(sweden)
summary(sweden)
}
\seealso{
\code{\link[BeePODYNA]{beepod_sweden}} for description of the dataset.
}
\keyword{datasets}
