% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_population.R, R/interaction.R
\name{as_population}
\alias{as_population}
\alias{as_population.default}
\alias{as_population.population}
\alias{as_population.list}
\alias{as_interaction.numeric}
\title{The fonction *as_population* takes as argument a list
containing the population caracteristics (population,
number at initial time, growth rates or birth rate and
death rate) or a poputaion monitored throught time.
And return the inoramtion as a list duable by BeePODyna package}
\usage{
as_population(object)

\method{as_population}{default}(object)

\method{as_population}{population}(object)

\method{as_population}{list}(object)

\method{as_interaction}{numeric}(object)
}
\arguments{
\item{object}{if it is list of populations objects (information on the population)
must be in the same semantic as population. It could be
a dataframe containing the evolution of the population throught time.}
}
\value{
If *object* is a list, the function will test if it match with the
 the requirements of a class *population* object.
 If *object* is a data frame the function will determine the growth rate,
 and form a list of class *population* containing the parameter of the population*.
}
\description{
The fonction *as_population* takes as argument a list
containing the population caracteristics (population,
number at initial time, growth rates or birth rate and
death rate) or a poputaion monitored throught time.
And return the inoramtion as a list duable by BeePODyna package
}
