% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_population.R
\name{as_population}
\alias{as_population}
\alias{as_population.default}
\alias{as_population.population}
\alias{as_population.list}
\title{as_population}
\usage{
as_population(object)

\method{as_population}{default}(object)

\method{as_population}{population}(object)

\method{as_population}{list}(object)
}
\arguments{
\item{object}{if it is list of populations objects (information on the population)}
}
\value{
If \code{'object'} is a list, the function will test if it match with the
 the requirements of a class \code{'population'} object.
 If \code{'object'} is a data frame the function will determine the growth rate,
 and form a list of class \code{'population'} containing the parameter of the population.
}
\description{
The fonction \code{'as_population'} takes as argument a list
containing the population caracteristics (population,
number at initial time, growth rates or birth rate and
death rate).
And return a list duable by \code{BeePODYNA} package
}
\examples{
obj = list(label ="pop.test",size= 100,growth_rate = 0.1,capacity = 500)
as_population(obj)

}
\author{
Martial
}
