% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community}
\alias{community}
\title{Create a community with giving its name, a minimum of one population.}
\usage{
community(label, population, ...)
}
\arguments{
\item{label}{the name of the community (character string)}

\item{population}{a population object}

\item{...}{additionnal populations objects. All other objects will be rejected and a warning will be print.}
}
\description{
Create a community with giving its name, a minimum of one population.
}
\examples{
hirsu = population("hirsuta",20,1,100)
daonen = population("daonensis",30,1.2,100)

daonen[[2]]=c(30,34)
daonen[[3]]=c(0,5)

alpha = community('alpha',hirsu,daonen)

}
\seealso{
\code{\link[BeePODYNA]{population}} to see how to make an object of class population
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
