% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{is.community}
\alias{is.community}
\title{is.community}
\usage{
is.community(x)
}
\arguments{
\item{x}{the object which must be a community to validate the condition}
}
\value{
a logical "TRUE" or "FALSE"
}
\description{
Check if the object is a community or not.
}
\examples{
hirsu = population("hirsuta",20,1,100)
daonen = population("daonensis",30,1.2,100)

daonen[[2]]=c(30,34)
daonen[[3]]=c(0,5)

alpha = community('alpha',hirsu,daonen)
is.community(alpha)

}
\seealso{
\code{\link[BeePODYNA]{community}} to see how to make an object of class community.
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
