% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{is.community}
\alias{is.community}
\title{Is.Community
The function determinines if the object is a population or not}
\usage{
is.community(x)
}
\arguments{
\item{x}{the object which must be a population to validate the condition}
}
\value{
a logical "TRUE" or "FALSE"
}
\description{
Is.Community
The function determinines if the object is a population or not
}
\author{
Jaunatre Maxime
}
