% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{logistic_growth}
\alias{logistic_growth}
\title{Build a logistic function for Verhulst population models}
\usage{
logistic_growth(pop, community, interactions)
}
\arguments{
\item{pop}{is the number of the population in the community on which apply the function. Giving the name of the pop as character is also possible.}

\item{community}{is an object of the class `community` on which to apply the function.}

\item{interactions}{is the object of the class `interactions` use to describe the interactions within the community.}
}
\description{
\code{logistic_growth} creates a fonction that can be used for an object of the class `beepodyna`.
}
\examples{
  data(hudson)
  species  = "hare"
  community = hudson
  int = interactions(2, c(-0.01,-0.09), labels=labels(hudson$populations))
  beepo_log = beepodyna(label = "hudson",
                        community = hudson,
                        interactions = int,
                        functions = logistic_growth,
                        verbose = TRUE)

  plot(simulate_n_pop_dynamic(beepo_log, 100)$community)

}
