% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{names_population}
\alias{names_population}
\title{names_population}
\usage{
names_population(pop)
}
\arguments{
\item{pop}{is an object of the class 'population'.}
}
\value{
It returns a character with the name of the population.
}
\description{
\code{names_population} returns the name of the population given in parameter.
}
\examples{
    pop<-population("Pop_1",42,2,200)
    names_population(pop)

}
\author{
Nicolas BARTALUCCI <bartalucci.nico@gmail.com>
}
