% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.population}
\alias{plot.population}
\title{plot.population}
\usage{
\method{plot}{population}(pop, capacity.line = F, log.time = F,
  log.pop = F, color = c("black", "red"), ...)
}
\arguments{
\item{capacity.line}{TRUE/FALSE if a dashed line must be drawned for each populations}

\item{log.time}{TRUE/FALSE if time axis (x axis) must be logarithmic}

\item{log.pop}{TRUE/FALSE if population axis (y axis) must be logarithmic}

\item{color}{a color vector for each population}

\item{...}{every argument for the \code{\link[graphics]{plot}} function such as graphical parameters for labels. Does NOT apply for lines and capacity ablines.}

\item{list}{a list of populations objects}
}
\description{
plot one or multiple given populations
}
\examples{
hirsu = population("hirsuta",20,1,100)
daonen = population("daonensis",30,1.2,100)

daonen[[2]]=c(30,34)
daonen[[3]]=c(0,5)

liste = list(hirsu,daonen)
plot_population(list = liste, log.pop = T, color = c(3,2), xlab = "Time")

}
\seealso{
\code{\link[graphics]{plot}} and \code{\link[graphics]{par}} for graphical parameter that were omitted on this documentation
}
\author{
Jaunatre Maxime
}
