% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_population}
\alias{plot_population}
\title{plot_population}
\usage{
plot_population(list = population("Primula", 1, 1, 100),
  capacity.line = F, log.time = F, log.pop = F, color = c("black",
  "red"), ...)
}
\arguments{
\item{list}{a list of populations objects}

\item{capacity.line}{TRUE/FALSE if a dashed line must be drawned for each populations}

\item{log.time}{TRUE/FALSE if time axis (x axis) must be logarithmic}

\item{log.pop}{TRUE/FALSE if population axis (y axis) must be logarithmic}
}
\description{
plot one or multiple given populations
}
\examples{
hirsu = population("hirsuta",20,1,100)
daonen = population("daonensis",30,1.2,100)

daonen[[2]]=c(30,34)
daonen[[3]]=c(0,5)

liste = list(hirsu,daonen)
plot.population(list = liste)

}
\author{
Jaunatre Maxime
}
