% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.population}
\alias{plot.population}
\title{plot.population}
\usage{
\method{plot}{population}(x, ...)
}
\arguments{
\item{x}{a population object}

\item{...}{every argument for the \code{\link[graphics]{plot}} function such as graphical parameters for lines. Does NOT apply for lines and capacity ablines.
\itemize{
  \item \strong{capacity_line} set to \code{TRUE} by default, if a dashed line must be drawned to show the capacity.
  \item \strong{capacity_lty} set to \code{3} by default.
  \item \strong{type} for the size line, set to \code{'b'} by default.
  \item \strong{xlab} set to \code{'Time'} by default, see \code{\link[graphics]{title}}
  \item \strong{ylab} set to \code{'Size'} by default, see \code{\link[graphics]{title}}
  \item \strong{main} set to use the pop label by default, see \code{\link[graphics]{title}}
  }}
}
\description{
plot one given population and its capacity line.
}
\examples{
hare = population("hare",30,1.2,33)
hare[[2]]=c(4,30,32,33,33)
hare[[3]]=c(0,1,2.5,5,8)

plot(hare)
plot(hare, col = 'red', pch=15, lty = 6, capacity_lty = 2)
plot(hare, col = 'red', pch=15, lty = 6, capacity_line = FALSE)

}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[graphics]{title}} and \code{\link[graphics]{par}} for plot parameter that were omitted on this documentation
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
