% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_population.R
\name{summary.population}
\alias{summary.population}
\title{Populations summaries and trends
#' @author Julia Guerra <jguerra_1995@hotmail.es>}
\usage{
\method{summary}{population}(object, ...,
  d.print = as.numeric(getOption("digits")))
}
\arguments{
\item{object}{an object of class "population"}

\item{...}{not implemented yet}

\item{d.print}{the number of decimal digits to be printed in statistics}
}
\value{
A list containing calculated summaries for a population. See 'examples'.
}
\description{
This function will allow the user to easily read
trends and characteristics (label, population size(s), growth rate(s)...) of a
population-class object.
}
\examples{
  # Default usage
  pop1 = population("mypop", 1, 0.8, 500)
  summary(pop1)

  # long history populations
  data(hudson)
  pop1 = hudson$hare
  sumrs1 = summary(pop1)   #for saving summaries list

  sumrs1$Rates_of_change        # taking a look into last growth tendencies

  # In case you want to modify digits printing settings

  summary(pop1,11)
  summary(pop1,getOption("digits")+2)

}
