% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.population.R
\name{summary.population}
\alias{summary.population}
\title{Population summaries and trends}
\usage{
\method{summary}{population}(populations,
  d.printing = getOption("digits"))
}
\arguments{
\item{d.printing}{the number of decimal digits to be printed in statistics}

\item{pop}{an object of class "population"}
}
\description{
This function will allow the user to easily read
trends and characteristics (label, population size(s), growth rate(s)...) of one or several
population-class objects. When dealing with a list of population-type objects, il will
recursively print the summaries for each one of them.
}
\examples{
# Default usage
pop = population("mypop", 1, 0.8, 500)
summary.population(pop)



# Several generations and population lists
pop1 = population("eva", 1, 0.8, 500)
pop2 = population("teo", 1, 0.7, 850)
pop2$size = c(1,50,750)

pops = list(pop1, pop2)

summary.population(pop2)
summary.population(pops)



# In case you want to modify digits' printing settings

summary.population(pop,11)
summary.population(pop,getOption("digits")+2)


}
